+++ Procfile
@@ -1,0 +1,2 @@
+web: npm start
+release: node scripts/migrate_db.js
