+++ index.html
@@ -17,24 +17,24 @@
               <thead>
               <th>Branch</th>
               <th>Development</th>
-              <th>Unit Tests Outcome</th>
+              <th>Status</th>
               <th>Test</th>
-              <th>Test testing outcome</th>
+              <th>Status</th>
               <th>Stage</th>
-              <th>Stage testing outcome</th>
+              <th>Status</th>
               <th>Production</th>
-              <th>Sanity Checks outcome</th>
+              <th>Status</th>
               </thead>
               <tbody>
                 <td>Master</td>
                 <td><a class="btn btn-primary">GO</a></td>
-                <td>100%</td>
+                <td>Green</td>
                 <td><a class="btn btn-primary">GO</a></td>
-                <td>something for Manual Testing confirmation</td>
+                <td>Green</td>
                 <td><a class="btn btn-primary">GO</a></td>
-                <td>Again testing format</td>
+                <td>Green</td>
                 <td><a class="btn btn-primary">GO</a></td>
-                <td>Sanity check docs / link</td>
+                <td>Green</td>
               </tbody>
               </table>
             <div id="app"></div>
