+++ results/last_build.env
@@ -1,0 +1,8 @@
+pkg_origin=echohack
+pkg_name=mint-dashboard
+pkg_version=0.1.0
+pkg_release=20171116042215
+pkg_ident=echohack/mint-dashboard/0.1.0/20171116042215
+pkg_artifact=echohack-mint-dashboard-0.1.0-20171116042215-x86_64-linux.hart
+pkg_sha256sum=6a209e7de7737a20f6d856992c1d7acf9c28b467b09f730d3f1eed4c95194708
+pkg_blake2bsum=9a5d3962909cadf8f31f7e724d0d82663adcf53aaaf5ab43980d3d835dd9846f
