+++ results/last_docker_export.env
@@ -1,0 +1,4 @@
+id=d1bbf71a6b36
+name=echohack/mint-dashboard
+tags=0.1.0-20171116042215,0.1.0,latest
+name_tags=echohack/mint-dashboard:0.1.0-20171116042215,echohack/mint-dashboard:0.1.0,echohack/mint-dashboard:latest
